Because we dont have our backend set up yet, I am testing this component using console.log. I add more complexity to my Login page through submit button which makes API call.
The second step for the assignment, I was able to test the console logs if a user enters their information.

The first test is to verify if data is log in. the second test is to verify if the email is invalid. This is not really possible because the form is validated but I still tested it.
The 3rd is to check if the password is too short.